WIP Added coverage.py, added coverage collection to test.py
Now coverage information can be collected if you provide the --coverage
to test.py. Internally this uses GCC's gcov instrumentation along with a
new script, coverage.py, that can parse *.gcov files.

This is mainly useful for finding coverage information in CI, and
because there is a risk that the instrumentation may affect debugging,
decided to not enable coverage by default.